# swagger.yml
swagger: "2.0"
info:
  title: "Star Supermarket"
  description: "Endpoints for the Star Supermarket System \n\n Team:\n- Carranza Torrico Juan Manuel\n- Paz Sea Yamil Ignacio\n- Saavedra Ayarde Stephanie Belén \n- Salazar Paredes Sara Oneydi"
  version: "1.0.0"
host: "api.star.supermarket"
basePath: "/v1"
paths:
...
  /product:
    get:
      tags:
      - "Product"
      summary: "Obtener todos los productos"
      description: "Tener la informacion de todos los productos"
      operationId: "getProductsList"
      responses:
        200:
          description: "Successful request - Retrieving Product List"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
  
    post:
      tags:
      - "Product"
      summary: "Create new Product"
      description: "This can only be done by the logged into the system."
      operationId: "createProduct"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created Product object"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "Successful request - User added"
        405:
          description: "Error while trying to add user"   

    put:
      tags:
      - "Product"
      summary: "Product User"
  
      description: "Allows you to update one of the following fields in an Product. One should be updated at a time"
      operationId: "updateUser"
      
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Update Product object"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "Successful request - User added"
        405:
          description: "Error while trying to add user"  

    delete:
        tags:
        - "Product"
        summary: "Delete a Product"
        description: "Delete a Product indicating his ID"
        operationId: "deleteProduct"
        produces:
        - "application/json"
        parameters:
        - in: "body"
          name: "body"
          description: "Update Product object"
          required: true
          schema:
            $ref: "#/definitions/Product"
        responses:
          200:
            description: "Successful request - Product added"
          405:
            description: "Error while trying to add Product"  
# Provides Swagger
  /provider:
    get:
      tags:
      - "Provider"
      summary: "Obtener todos los Provider"
      description: "Tener la informacion de todos los Provider"
      operationId: "getProvidersList"
      responses:
        200:
          description: "Successful request - Retrieving Provider List"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Provider"

    post:
      tags:
      - "Provider"
      summary: "Create new Provider"
      description: "This can only be done by the logged into the system."
      operationId: "createProvider"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created Provider object"
        required: true
        schema:
          $ref: "#/definitions/Provider"
      responses:
        200:
          description: "Successful request - Provideradded"
        405:
          description: "Error while trying to add Provider"   

    put:
      tags:
      - "Provider"
      summary: "Provider User"
  
      description: "Allows you to update one of the following fields in an Product. One should be updated at a time"
      operationId: "updateProvider"
      
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Update Provider object"
        required: true
        schema:
          $ref: "#/definitions/Provider"
      responses:
        200:
          description: "Successful request - User added"
        405:
          description: "Error while trying to add user"  



# User Swagger
  /user:
    get:
      tags:
      - "user"
      summary: "Get All Users"
      description: "Gets all currently registered users"
      operationId: "getUsersList"
      responses:
        200:
          description: "Successful request - Retrieving User List"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
    
    post:
      tags:
      - "user"
      summary: "Create new user"
      description: "This can only be done by the logged into the system."
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful request - User added"
        405:
          description: "Error while trying to add user"    
  /user/{userId}:
    get:
      tags:
      - "user"
      summary: Get user info by Id
      operationId: "getUserInfo"
      parameters:
      
      - in: "path"
        name: "userId"
        type: "integer"
        format: "int32"
        required: true
      
      responses:
        200:
          description: "Successful Request"
          schema:
            $ref: "#/definitions/User"
        404:
          description: "Error - User Does Not Exist"
   
    put:
      tags:
      - "user"
      summary: "Update User"
  
      description: "Allows you to update one of the following fields in an user. One should be updated at a time"
      operationId: "updateUser"
      
      parameters:
      - in: "path"
        name: "userId"
        description: "ID ofuser to be updated"
        type: "integer"
        format: "int32"
        required: true
            
      - in: "query"
        name: "field"
        description: "field to be updated"
        type: "string" 
        required: true
        
      - in: "query"
        name: "value"
        description: "New value of the previous field"
        type: "string"
        required: true
      
      responses:
        200:
          description: "Successful Request - User Updated"
        404:
          description: "Error - User Does Not Exist"
      
    delete:
      tags:
      - "user"
      summary: "Delete an user"
      description: "Delete an user indicating his ID"
      operationId: "deleteUser"
      
      parameters:
      - in: "path"
        name: "userId"
        type: "integer"
        format: "int32"
        required: true
        
      responses:
        200:
          description: "Successful Request - User Deleted"
        404:
          description: "Error - User Does Not Exist"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://star-supermarket/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "Modify fdata in your account"
      read:pets: "Read your data"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"

# Models (Definitions)
definitions:
...
  Product:
    type: "object"
    properties:
      id_producto:
        type: "string"
        description: "Es el id que permite poder identificar al producto mediante una llave"
      nombre:
        type: "string"
        description: "Nombre del producto"
      descripcion:
        type: "string"
        description: "Descripcion del producto"
      categoria:
        type: "string"
        description: "Categoria del producto"
      precio_venta:
        type: "number"
        format: "double"
        description: "Precio de venta"
      stock:
        type: "integer"
        description: "Stock del producto"
      precio_compra:
        type: "integer"
        format: "double"
        description: "Precio de compra con los proveedores"
      estado:
        type: "integer"
        description: "Estado de activo e inactivo" 
      fecha_vencimiento:
        type: "integer"
        format: "date"
        description: "Fecha de vencimiento del producto"
  Provider:
    type: "object"
    properties:
      id_proveedor:
        type: "string"
        description: "Identificador del usuario"
      name:
        type: "string"
        description: "Nombre del Proveedor"
      lastname:
        type: "string"
        description: "Apellido del Proveedor"
      company:
        type: "string"
        description: "Nombre del proveedor"
      email:
        type: "string"
        description: "Correo Electronico del proveedor"
      phone:
        type: "number"
        description: "Numero de telefono del proveedor"
      estado:
        type: "integer"
        description: "Estado de activo e inactivo"  
  User: 
    type: "object"
    properties:
      uid_usuario:
        type: "string"
        description: "Identificador del usuario"
      identificador:
        type: "string"
        description: "Identificador del usuarios"
      fecha_creacion:
        type: "string"
        format: "date-time"
        description: "Fecha de Creacion"
      fecha_acceso:
        type: "string"
        format: "date-time"
        description: "Fecha de Acceso"
...